requirements.txt
saving-place
beautifulsoup4>=4.5.3
websockets>=3.3
requests>=2.13
Pillow>=4.0.0
